move deallocate and volumeDelete to instance.remove
